pyproject.toml
[project]
name = "shdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "mcp[cli]>=1.12.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
